st_ocr/requirements.txt
# =========================
# STREAMLIT & UI
# =========================
streamlit==1.48.1

# =========================
# IMAGE PROCESSING
# =========================
numpy==1.26.4
pillow==11.3.0
opencv-python==4.6.0.66

# =========================
# OCR
# =========================
paddleocr==2.6.0
paddlepaddle==2.6.3
